[refactor] removing dataCollect class and transfering its method into VisitedPlacesComputer
